+++ spec/rspec/matchers/description_generation_spec.rb
@@ -127,7 +127,7 @@ describe "Matchers should be able to generate their own descriptions" do
 
   it "expect(array).to contain_exactly(1, 2, 3)" do
     expect([1,2,3]).to contain_exactly(1, 2, 3)
-    expect(RSpec::Matchers.generated_description).to eq "should contain exactly 1, 2 and 3"
+    expect(RSpec::Matchers.generated_description).to eq "should contain exactly 1, 2, and 3"
   end
 
   it "expect(...).to match" do